Added lots of stuff to the Design file
Design file is intended to be the primary source of our documentation, holding information about what the various files do. This will need to be expanded to include what you'd need to do in order to get the server running. Still have about half of the files to document, going to work more on the important files.