Added a hack solution to downloading data from the custom graph.
We want to be able to get the data from graphs when we look at the custom graph. As a temporary solution, the contents of the custom graph will be written to a data table. Clicking a button will display that table. VERY hack solution, but will at least see if this is feasible.