Fixed some formatting.
Removed the legend because its now redundant. Also flipped the orientation of the placeholder table in the custom graph so that it approaches readable. Unnecessary since I'm planning on removing the table in favor of xml, but imagine writing to xml requires a similar method, so can be salvaged for that.