Adding the current data reading column.
Requires making a query in views.py. This query must be parsed by urls.py to make sure we get the right input. This data then is sent to energy_statistics table's html then to the javascript page. BTW, should consider changing the URLs and naming to be consistent. statistics_table is a good name. Let's use it.